requirements.txt
streamlit
pdfplumber
pdf2image
pytesseract
PyYAML
Pillow
fuzzywuzzy[speedup]
python-Levenshtein
pytest
